requirements.txt
numpy==1.19.5
pandas==1.2.0
yfinance==0.1.55
python-dotenv==0.15.0
SQLAlchemy==1.3.22
psycopg2==2.8.6
requests==2.25.1
